Completed skeletons of both pages with code organization
The code organization of CSS, Images, scripts, etc.  is all here.  There
is a common CSS for both pages, to give the same look and feel, and a
div to hold the page specific content.  Note there is also a debug.css
(uncomment the HTML to use this)

RE: Reports -- I am using jQuery tabs, but now think it may be easier to
just fake it by replacing a fixed size div. I am also not too keen on
the absolute positioning of the dive for the elements.

I also have not yet done any .js behaviors.  The tab (or div) for each
neighborhood should be generated in .js

Comments welcome.